requirements.txt
# rasrss dependencies
flask>=3.0
feedparser>=6.0
requests>=2.31
openai>=1.0
python-dotenv>=1.0
apscheduler>=3.10
gitpython>=3.1
